Refactor: change Position type and related code to use uint16
Changed the `Position` type from `int` to `uint16` to improve memory efficiency and avoid negative coordinates. Updated all related code, including tests, to handle the new type. This includes changes to `Piece`, `Block`, `Playfield`, and `Tetromino` modules, as well as the addition of `FSharpPlus` package for non-empty set operations.

+++ Tetris.Core/Position.fs
@@ -5,7 +5,7 @@ namespace Tetris.Core
 /// Represents the position in a two-dimensional space.
 /// </summary>
 [<Struct>]
-type Position = { X: int; Y: int }
+type Position = { X: uint16; Y: uint16 }
 
 module Position =
     /// <summary>
